fix migration script
error was:
```
psql:C:/workspaces/metasfresh/metasfresh/backend/de.metas.fresh/de.metas.fresh.base/src/main/sql/postgresql/system/70-de.metas.fresh/5748890_sys_gh20312_Docs_Purchase_InOut_Details.sql:167: ERROR:  return type mismatch in function declared to return record
	DETAIL:  Final statement returns bigint instead of numeric at column 14.
	CONTEXT:  SQL function "docs_purchase_inout_details"

```

+++ backend/de.metas.fresh/de.metas.fresh.base/src/main/sql/postgresql/system/70-de.metas.fresh/5748890_sys_gh20312_Docs_Purchase_InOut_Details.sql
@@ -37,7 +37,7 @@ FROM (SELECT Attributes,
              iol.Description,
              bp_product_no,
              bp_product_name,
-             (ROW_NUMBER() OVER (ORDER BY MAX(iol.line))) * 10 AS line
+             ((ROW_NUMBER() OVER (ORDER BY MAX(iol.line))) * 10)::numeric AS line
 
       FROM
           -- Sub select to get all in out lines we need. They are in a subselect so we can neatly group by the attributes
